Split out localkube commands to interface
