+++ cmd/minikube/cmd/logs.go
@@ -21,9 +21,11 @@ import (
 	"log"
 	"os"
 
+	"github.com/golang/glog"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+	cmdcfg "k8s.io/minikube/cmd/minikube/cmd/config"
 	cmdUtil "k8s.io/minikube/cmd/util"
-	"k8s.io/minikube/pkg/minikube/cluster"
 	"k8s.io/minikube/pkg/minikube/machine"
 )
 
@@ -43,7 +45,13 @@ var logsCmd = &cobra.Command{
 			os.Exit(1)
 		}
 		defer api.Close()
-		s, err := cluster.GetHostLogs(api, follow)
+
+		clusterBootstrapper, err := GetClusterBootstrapper(api, viper.GetString(cmdcfg.Bootstrapper))
+		if err != nil {
+			glog.Exitf("Error getting cluster bootstrapper: %s", err)
+		}
+
+		s, err := clusterBootstrapper.GetClusterLogs(follow)
 		if err != nil {
 			log.Println("Error getting machine logs:", err)
 			cmdUtil.MaybeReportErrorAndExit(err)
