+++ cmd/minikube/cmd/status.go
@@ -24,6 +24,8 @@ import (
 	"github.com/docker/machine/libmachine/state"
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+	cmdcfg "k8s.io/minikube/cmd/minikube/cmd/config"
 	cmdUtil "k8s.io/minikube/cmd/util"
 	"k8s.io/minikube/pkg/minikube/cluster"
 	"k8s.io/minikube/pkg/minikube/config"
@@ -62,9 +64,13 @@ var statusCmd = &cobra.Command{
 		ls := state.None.String()
 		ks := state.None.String()
 		if ms == state.Running.String() {
-			ls, err = cluster.GetLocalkubeStatus(api)
+			clusterBootstrapper, err := GetClusterBootstrapper(api, viper.GetString(cmdcfg.Bootstrapper))
 			if err != nil {
-				glog.Errorln("Error localkube status:", err)
+				glog.Exitf("Error getting cluster bootstrapper: %s", err)
+			}
+			ls, err = clusterBootstrapper.GetClusterStatus()
+			if err != nil {
+				glog.Errorln("Error cluster status:", err)
 				cmdUtil.MaybeReportErrorAndExit(err)
 			}
 			ip, err := cluster.GetHostDriverIP(api)
